test: added tests for testing connectivity over nginx as a proxy
- changed compile-and-test.sh scripts, so it starts both standard and "proxy"
- added simple nginx config for docker container
- tests related to UDP communication moved to separate class and skipped in case of http proxy testing
